ref #8 Python cleanup
- Move model base classes out of \__init\__.py.
- Extend __object__ in base classes.
- Convert some of the doc styling to sphinx style.
